Declare module as github.com/gojek/work (#16)
Earlier, the module was meant to be used with a replace directive for
github.com/gocraft/work. Declaring it as a separate module makes it
easier to build wrapper libraries as replace directives always need to
be specified in the main module. If the library uses a feature not
present in gocraft/work, it would result in hard-to-follow compilation
errors:

	# source.company.io/mygroup/worker-adapter
	../worker-adapter/adapter.go:289:89: queue.MaxConcurrency undefined (type *work.Queue has no field or method MaxConcurrency)
	../worker-adapter/adapter.go:290:84: queue.LockCount undefined (type *work.Queue has no field or method LockCount)

A somewhat better alternative is to declare as a separate module which
would result in easier to fix compilation errors.